Change search index presenter to indexer
This reduces the coupling between the Policy models and rummager.
